+++ web/src/lib/components/shared-components/gallery-viewer/gallery-viewer.svelte
@@ -23,7 +23,7 @@
   import { debounce } from 'lodash-es';
   import { t } from 'svelte-i18n';
   import AssetViewer from '../../asset-viewer/asset-viewer.svelte';
-  import DeleteAssetDialog from '../../photos-page/delete-asset-dialog.svelte';
+  import DeleteAssetDialog from '../../timeline-viewer/actions/delete-asset-dialog.svelte';
   import Portal from '../portal/portal.svelte';
 
   interface Props {
