+++ web/src/lib/components/timeline-viewer/actions/delete-asset-dialog.svelte
